move files from 'trainingentities' package to 'training' package
