+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/challenge/ChallengeUiState.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.presentation.screens.trainingentities.challenge
+package com.google.jetfit.presentation.screens.training.challenge
 
 data class ChallengeUiState(
     val isLoading: Boolean = true,
