+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/challenge/ChallengeViewModel.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.presentation.screens.trainingentities.challenge
+package com.google.jetfit.presentation.screens.training.challenge
 
 import androidx.lifecycle.ViewModel
 import com.google.jetfit.data.repositories.JetFitRepository
