+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/composables/RegimenDescription.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.presentation.screens.trainingentities.composables
+package com.google.jetfit.presentation.screens.training.composables
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
