+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/series/SeriesViewModel.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.presentation.screens.trainingentities.routine
+package com.google.jetfit.presentation.screens.training.series
 
 import androidx.lifecycle.ViewModel
 import com.google.jetfit.data.repositories.JetFitRepository
@@ -9,31 +9,33 @@ import kotlinx.coroutines.flow.update
 import javax.inject.Inject
 
 @HiltViewModel
-class RoutineViewModel @Inject constructor(
+class SeriesViewModel @Inject constructor(
     repository: JetFitRepository
 ) : ViewModel() {
     val id: String = "1"
 
-    private val _state: MutableStateFlow<RoutineUiState> by lazy {
+    private val _state: MutableStateFlow<SeriesUiState> by lazy {
         MutableStateFlow(
-            RoutineUiState()
+            SeriesUiState()
         )
     }
     val state = _state.asStateFlow()
 
     init {
         try {
-            repository.getRoutineById(id).also { routine ->
+            repository.getSeriesById(id).also { series ->
                 _state.update {
                     it.copy(
                         isLoading = false,
                         error = null,
-                        subtitle = "${routine.instructorName}  |  ${routine.workoutType.value}",
-                        title = routine.name,
-                        description = routine.description,
-                        duration = "${routine.duration} min",
-                        intensity = routine.intensity.value,
-                        imageUrl = routine.imageUrl
+                        subtitle = "${series.instructorName}  |  ${series.intensity.value}",
+                        title = series.name,
+                        description = series.description,
+                        intensity = series.intensity.value,
+                        numberOfClasses = series.numberOfClasses.toString(),
+                        numberOfWeeks = series.numberOfWeeks.toString(),
+                        minutesPerDay = series.minutesPerDay.toString(),
+                        imageUrl = series.imageUrl
                     )
                 }
             }
